multiplication & stuff

+++ src/main/kotlin/ru/spbstu/logic/lib/Reify.kt
@@ -5,15 +5,16 @@ import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.plus
 import ru.spbstu.logic.App
 import ru.spbstu.logic.Expr
+import java.math.BigInteger
 
 fun reify(expr: Expr): Any? = when (expr) {
-    natZero -> 0
+    natZero -> 0.toBigInteger()
     nil -> persistentListOf<Expr>()
     is App -> when (expr.f) {
         natBits -> try {
-            val base = reify(expr.args[1]) as Int
-            val cc = if (expr.args[0] == oneBit) 1 else 0
-            cc + base * 2
+            val base = reify(expr.args[1]) as BigInteger
+            val cc = if (expr.args[0] == oneBit) 1.toBigInteger() else 0.toBigInteger()
+            cc + base * 2.toBigInteger()
         } catch (e: Exception) {
             expr.toString()
         }
